requirements.txt
aiogram
aiohttp
tortoise-orm[asyncpg]
rich
uvloop
ujson
python-rapidjson
aerich
validators